Add --cleanup processing and --verbose output to the C test suite. This lets fs-test and repos-test clean up after themselves.
Add a machanism for creating new working copy and repository paths
(and --cleanup processing for same) to the Python tests.

Enhance the Windows tests so that they can run outside the source
tree (for example, on a RAM-backed volume).


* subversion/include/svn_test.h
 (svn_test_add_dir_cleanup): New prototype.

* subversion/tests/svn_tests_main.c
  (verbose_mode, cleanup_mode) New flags.
  (cleanup_pool): New global pool for registering cleanups.
  (cleanup_rmtree, svn_test_add_dir_cleanup): New functions.
  (main): Scan the argument list for --cleanup and --verbose flags, and
  initialize cleanup_mode and verbose_mode accordingly.
  Create and clear cleanup_pool in lockstep with test_pool.

* subversion/tests/fs-helpers.c
  (svn_test__create_fs, svn_test__create_repo): Register the path of
  the new file system or repository for cleanup.

* build/run_tests.py
 (TestHarness.run): Pass the --verbose flag to C tests.

* subversion/tests/python-libs/exectest.py
  (run_tests): Pass all command line arguments to the sub-tests.

* subversion/tests/clients/cmdline/svntest/main.py
  (Sandbox.__init__): Set self.repo_url to the URL of self.repo_dir.
  (Sandbox.add_test_path): New parameter 'remove'; when non-zero (which is
  the default), delete the directory tree at 'path'.
  (Sandbox.add_repo_path, Sandbox.add_wc_path): New. Create WC or repo
  paths for the tests, and register them for cleanup.

* subversion/tests/clients/cmdline/copy_tests.py,
  subversion/tests/clients/cmdline/stat_tests.py,
  subversion/tests/clients/cmdline/svnversion_tests.py,
  subversion/tests/clients/cmdline/prop_tests.py,
  subversion/tests/clients/cmdline/basic_tests.py,
  subversion/tests/clients/cmdline/update_tests.py,
  subversion/tests/clients/cmdline/trans_tests.py,
  subversion/tests/clients/cmdline/commit_tests.py,
  subversion/tests/clients/cmdline/merge_tests.py,
  subversion/tests/clients/cmdline/switch_tests.py,
  subversion/tests/clients/cmdline/externals_tests.py: Use Sandbox.add_wc_path
  and Sandbox.add_repo_path to crate new test paths, and use sbox.repo_url
  instead of calculating.

* subversion/tests/clients/cmdline/stat_tests.py
  (status_uninvited_parent_directory): Adjust the regex to the new WC name.

* subversion/tests/clients/cmdline/externals_tests.py
  (externals_test_cleanup): Gone. Superseded by the new Sandbox code.

* win-tests.py: Accept an optional target dir for running the tests, then
  copy the test programs there and chdir to it before running the tests.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@845430 13f79535-47bb-0310-9956-ffa450edef68

+++ win-tests.py
@@ -46,8 +46,8 @@ import getopt
 
 opts, args = getopt.getopt(sys.argv[1:], 'rdvcu:',
                            ['release', 'debug', 'verbose', 'cleanup', 'url='])
-if len(args):
-  print 'Warning: non-option arguments will be ignored'
+if len(args) > 1:
+  print 'Warning: non-option arguments after the first one will be ignored'
 
 # Interpret the options and set parameters
 all_tests = tests + fs_tests + client_tests
@@ -66,7 +66,7 @@ for opt,arg in opts:
   elif opt in ['-v', '--verbose']:
     verbose = 1
   elif opt in ['-c', '--cleanup']:
-    cleanup=1
+    cleanup = 1
   elif opt in ['-u', '--url']:
     all_tests = client_tests
     repo_loc = 'remote repository ' + arg + '.'
@@ -75,15 +75,40 @@ for opt,arg in opts:
 
 print 'Testing', filter, 'configuration on', repo_loc
 
+# Calculate the source and test directory names
+abs_srcdir = os.path.abspath("")
+if len(args) == 0:
+  abs_builddir = abs_srcdir
+  remove_execs = 1
+  create_dirs = 0
+else:
+  abs_builddir = os.path.abspath(args[0])
+  remove_execs = 0
+  create_dirs = 1
+
+
 # Have to move the executables where the tests expect them to be
 copied_execs = []   # Store copied exec files to avoid the final dir scan
+
+def create_target_dir(dirname):
+  if create_dirs:
+    tgt_dir = os.path.join(abs_builddir, dirname)
+    if not os.path.exists(tgt_dir):
+      if verbose:
+        print "mkdir:", tgt_dir
+      os.makedirs(tgt_dir)
+
 def copy_execs(filter, dirname, names):
   global copied_execs
-  if os.path.basename(dirname) != filter: return
+  if os.path.basename(dirname) != filter:
+    return
   for name in names:
-    if os.path.splitext(name)[1] != ".exe": continue
+    if os.path.splitext(name)[1] != ".exe":
+      continue
     src = os.path.join(dirname, name)
-    tgt = os.path.join(os.path.dirname(dirname), name)
+    dir = os.path.dirname(dirname)
+    tgt = os.path.join(abs_builddir, dir, name)
+    create_target_dir(dir)
     try:
       if verbose:
         print "copy:", src
@@ -94,30 +119,36 @@ def copy_execs(filter, dirname, names):
       traceback.print_exc(file=sys.stdout)
       pass
 os.path.walk("subversion", copy_execs, filter)
-
+create_target_dir('subversion/tests/clients/cmdline')
 
 # Run the tests
-abs_srcdir = os.path.abspath("")
-abs_builddir = abs_srcdir  ### For now ...
-
 sys.path.insert(0, os.path.join(abs_srcdir, 'build'))
 import run_tests
 th = run_tests.TestHarness(abs_srcdir, abs_builddir, sys.executable, None,
-                           os.path.abspath(log),
-                           base_url=base_url, verbose=1, cleanup=cleanup)
-failed = th.run(all_tests)
+                           os.path.join(abs_builddir, log),
+                           base_url, 1, cleanup)
+old_cwd = os.getcwd()
+try:
+  os.chdir(abs_builddir)
+  failed = th.run(all_tests)
+except:
+  os.chdir(old_cwd)
+  raise
+else:
+  os.chdir(old_cwd)
 
 
 # Remove the execs again
-for tgt in copied_execs:
-  try:
-    if os.path.isfile(tgt):
-      if verbose:
-        print "kill:", tgt
-      os.unlink(tgt)
-  except:
-    traceback.print_exc(file=sys.stdout)
-    pass
+if remove_execs:
+  for tgt in copied_execs:
+    try:
+      if os.path.isfile(tgt):
+        if verbose:
+          print "kill:", tgt
+        os.unlink(tgt)
+    except:
+      traceback.print_exc(file=sys.stdout)
+      pass
 
 
 # Print final status
